utilities: ovs-ctl.in: Allow passing DPDK options to ovs-vswitchd
The ovs-ctl script is used to launch ovs-vswitchd among other things.
However it does not make it possible to pass DPDK options to the
daemon. We fix this by explicitly looking and extracting the DPDK
options from the command line which is then reconstructed so it can be
parsed by ovs-ctl as usual. The DPDK options are later passed to
ovs-vswitchd.

Signed-off-by: Markos Chandras <mchandras@suse.de>